Fix irace start code and windows support

+++ R/app.R
@@ -281,7 +281,9 @@ App <- R6::R6Class(
 
       unlink(file.path(gui$optionsPath, ".Fimages"), recursive = TRUE, force = TRUE)
       unlink(file.path(gui$optionsPath, ".Pimages"), recursive = TRUE, force = TRUE)
-      unlink(pkg$tempFolder, recursive = TRUE, force = TRUE)
+      if (!get_option("debug", FALSE)) {
+        unlink(pkg$tempFolder, recursive = TRUE, force = TRUE)
+      }
     }
   )
 )
